feat: wIP: Refactor of calendar controls started
