+++ module/src/hooks/calendar/index.ts
@@ -1,0 +1,9 @@
+/** ***************************************
+ * CALENDAR - Barrel file.
+ * Exports everything calendar related from the individual files
+ * This file is then imported into the root index and exported under the "Form" namespace.
+ **************************************** */
+
+export * from './calendar.hooks';
+export * from './calendar.types';
+export * from './calendar.utils';
